struct type

+++ src/main/java/org/glavo/llvm/internal/ir/CompositeTypeImpl.java
@@ -13,5 +13,5 @@ public final class CompositeTypeImpl extends NativeImpl {
 
     //todo: public static native boolean indexValid(@Handle(CompositeType.class) int handle, @Handle(Value.class) long v);
 
-    public static native boolean indexValid(@Handle(CompositeType.class) int handle, @Unsigned int idx);
+    public static native boolean indexValid(@Handle(CompositeType.class) long handle, @Unsigned int idx);
 }
